feat: dir Refactoring
