+++ lib/main.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 
 import 'package:bitsdojo_window/bitsdojo_window.dart';
 import 'package:desktop_multi_window/desktop_multi_window.dart';
-import 'package:fl_caption/common/rust/candle_models/whisper.dart' show WhisperStatus;
+import 'package:fl_caption/common/rust/candle_models/whisper/model.dart' show WhisperStatus;
 import 'package:fl_caption/common/translate/translate_provider.dart';
 import 'package:fl_caption/pages/settings/settings_page.dart';
 import 'package:fl_caption/widgets/error.dart';
