+++ rust/src/candle_models/whisper/assets/whisper/melfilters.bytes
