+++ rust/src/candle_models/whisper/model.rs
